// client/src/features/auth/type.ts
export type AuthMode = 'login' | 'register';

export interface AuthFormData {
  fullname?: string;
  email: string;
  password: string;
  confirmPassword?: string;
}

export interface AuthFormProps {
  mode: AuthMode;
  onSubmit: (data: AuthFormData) => void;
  onModeChange: (mode: AuthMode) => void;
}